# Speed up the Haskell implementation