--- a/haskell/Makefile
+++ b/haskell/Makefile
@@ -1,8 +1,13 @@
+LLC := llc
+OPT := opt
+
 v1:
-	ghc -O2 main.hs
+	ghcup set ghc 8.10.7
+	cabal build -O2 --ghc-options="-pgmlo=${OPT} -pgmlc=${LLC}"
+	cp $(shell cabal list-bin -O2 exe:haskell) ./main
 
 strip:
 	strip -s main
 
 clean:
-	rm main main.o main.hi
+	rm -rf main dist-newstyle