--- a/haskell/haskell.cabal
+++ b/haskell/haskell.cabal
@@ -1,0 +1,10 @@
+cabal-version:      2.4
+name:               haskell
+version:            0.1.0.0
+
+executable haskell
+    main-is:          main.hs
+    build-depends:    base
+    ghc-options:      -Wall
+    hs-source-dirs:   .
+    default-language: Haskell2010