--- a/haskell/haskell.cabal
+++ b/haskell/haskell.cabal
@@ -5,6 +5,6 @@ version:            0.1.0.0
 executable haskell
     main-is:          main.hs
     build-depends:    base, vector
-    ghc-options:      -Wall
+    ghc-options:      -Wall -fllvm
     hs-source-dirs:   .
     default-language: Haskell2010