--- a/haskell/haskell.cabal
+++ b/haskell/haskell.cabal
@@ -4,7 +4,7 @@ version:            0.1.0.0
 
 executable haskell
     main-is:          main.hs
-    build-depends:    base
+    build-depends:    base, vector
     ghc-options:      -Wall
     hs-source-dirs:   .
     default-language: Haskell2010