This PR speeds up the Haskell implementation significantly using the LLVM backend.

Yesterday I saw someone wondering why the Haskell implementation was significantly slower than the other implementations in this test, so I took a stab at optimizing it.

The key points are:

* Use unboxed vectors instead of (singly linked) lists
* Replace the div/mod functions, which are known to be slow, with quot/rem
* Switch from the native code generator to the LLVM backend, which is known to generate faster code at expense of slower compilation

Here's the result on my MacBook Pro 16'':

```
% hyperfine {c,cpp,rust,haskell}/main
Benchmark #1: c/main
  Time (mean ± σ):      2.327 s ±  0.051 s    [User: 2.302 s, System: 0.018 s]
  Range (min … max):    2.272 s …  2.418 s    10 runs

Benchmark #2: cpp/main
  Time (mean ± σ):      3.556 s ±  0.069 s    [User: 3.519 s, System: 0.028 s]
  Range (min … max):    3.470 s …  3.688 s    10 runs

Benchmark #3: rust/main
  Time (mean ± σ):      2.633 s ±  0.079 s    [User: 2.584 s, System: 0.019 s]
  Range (min … max):    2.582 s …  2.848 s    10 runs

Benchmark #4: haskell/main
  Time (mean ± σ):      3.105 s ±  0.058 s    [User: 3.049 s, System: 0.028 s]
  Range (min … max):    3.064 s …  3.264 s    10 runs

Summary
  'c/main' ran
    1.13 ± 0.04 times faster than 'rust/main'
    1.33 ± 0.04 times faster than 'haskell/main'
    1.53 ± 0.04 times faster than 'cpp/main'
```

System prerequisites are:

* LLVM >= 9 && < 13
* GHC 8.10.7 and cabal-install are installed using [ghcup](https://www.haskell.org/ghcup/)
    * Install ghcup then `ghcup install ghc 8.10.7 && ghcup install cabal latest`

As noted in the [commit message](https://github.com/jabbalaci/SpeedTests/commit/5b6d792239cca566ae78dbcf2f8da2e167a76b59), the Makefile expects that llc and opt are available in $PATH. If not please set OPT and LLC accordingly when running make.