@jabbalaci 2021-09-26
Thanks. Nice optimization! Please check the dir. `haskell/`, I made some changes. I changed the `any.base ==4.14.3.0` to `any.base ==4.14.2.0` for instance, otherwise the compilation failed on my system.