@maoe 2021-09-26
I guess that's because you're probably not using GHC 8.10.7. If that change works for you, I think it's fine.