@maoe 2021-09-27
Ah apparently you're using 8.10.5 according to the updated README. In that case we should update https://github.com/jabbalaci/SpeedTests/blob/d4090d83590b50b7d6b7b26f1b8c2b6929773a03/haskell/Makefile#L13 accordingly.