@jabbalaci 2021-09-27
Found it. In the PATH, I had to put `~/.ghcup/bin` before `/bin`, and now it finds GHC 8.10.7 . I restored `any.base ==4.14.3.0` and re-run the tests. There's no change in the runtime but the EXE is much bigger than the EXE produced by 8.10.5 . Thanks for the contribution!